test: Convert `test_types` integration tests to unit tests
The "integration" tests in `redis/tests/test_types.rs` are actually unit
tests. So we turn them into unit tests.

This is part of #2249
